rearrange some of the code
