+++ src/factories/service_processor_factory.py
@@ -1,24 +1,13 @@
+import os
+
 from rococo.messaging import BaseServiceProcessor
-import logging
-
-# Configure logging
-logging.basicConfig(level=logging.INFO)
-
-
-# this will go elsewhere
-class LoggingServiceProcessor(BaseServiceProcessor):
-
-    def process(self, message):
-        logging.info(f"Received message: {message}")
-
-    def __init__(self):
-        super().__init__()
+from service_processors.service_processors import LoggingServiceProcessor
 
 
 def get_service_processor() -> BaseServiceProcessor:
     service_processor = BaseServiceProcessor()
-    # hard coding for now
-    if True:
+    processor_type = os.environ.get("PROCESSOR_TYPE")
+    if processor_type == "LoggingServiceProcessor":
         # if condition for LoggingServiceProcessor, then:
         service_processor = LoggingServiceProcessor()
     return service_processor
