+++ src/service_processors/__init__.py
@@ -1,0 +1,1 @@
+from .service_processors import LoggingServiceProcessor
